Framework: refactor CartStore
