+++ client/lib/store-transactions/free-transactions.js
@@ -2,10 +2,10 @@
  * Internal dependencies
  */
 var wpcom = require( 'lib/wp' ).undocumented(),
+	cartLib = require( 'lib/cart' ),
 	cartValues = require( 'lib/cart-values' ),
 	cartItems = cartValues.cartItems,
-	plans = require( 'lib/plans-list' )(),
-	productsList = require( 'lib/products-list' )();
+	plans = require( 'lib/plans-list' )();
 
 export function freeTrial( planName, onComplete ) {
 	const emptyCart = cartValues.emptyTemporaryCart( this.props.site.ID ),
@@ -17,7 +17,7 @@ export function freeTrial( planName, onComplete ) {
 
 export function submitFreeTransaction( cart, onComplete ) {
 	const transaction = {
-		cart: cartValues.fillInAllCartItemAttributes( cart, productsList.get() ),
+		cart: cartLib.fillInAllCartItemAttributes( cart ),
 		payment: { paymentMethod: 'WPCOM_Billing_WPCOM' }
 	};
